Redefined movement from one method to 2
changed m_manipulatorMotor.set(speed);
to
m_manipulatorMotor.set(ControlMode.PercentOutput, speed);
Speed is a variable that is passed in when the method is called.
Added a method that stopped the manipulator.
Defined movement directions for the actions, they will either work correctly or be reversed. If they are reversed then we can setInverted to true.
I highly doubt that this will negatively affect the code Isaac.